chore: common.css 및 자잘한 코드 수정

+++ src/main/java/com/kl/grooveo/boundedContext/comment/controller/FreedomPostCommentController.java
@@ -12,10 +12,13 @@ import com.kl.grooveo.boundedContext.member.service.MemberService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 @RequestMapping("/comment")
 @RequiredArgsConstructor
@@ -26,6 +29,7 @@ public class FreedomPostCommentController {
     private final MemberService memberService;
     private final Rq rq;
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/create/{id}")
     public String create(Model model, @PathVariable("id") Long id,
                          @RequestParam(value = "commentPage", defaultValue = "0") int commentPage,
@@ -47,12 +51,15 @@ public class FreedomPostCommentController {
                 freedomPostComment.getFreedomPost().getId(), freedomPostComment.getId());
     }
 
+    @PreAuthorize("isAuthenticated()")
     @DeleteMapping("/{id}")
     public String delete(@PathVariable("id") Long id) {
         FreedomPostComment freedomPostComment = this.freedomPostCommentService.getComment(id);
-//        if (!freedomPostComment.getAuthor().getUsername().equals(principal.getName())) {
-//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
-//        }
+
+        if (!freedomPostComment.getAuthor().getUsername().equals(rq.getMember().getUsername())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+        }
+
         this.freedomPostCommentService.delete(freedomPostComment);
         return String.format("redirect:/community/freedomPost/detail/%s", freedomPostComment.getFreedomPost().getId());
     }
